cardano-node:  factor tracing-independent mini-API as Cardano.Node.Queries.*
The current tracing system makes a number of projections of the system state,
which are not really tracing-specific.

We consolidate those as modules:

  - Cardano.Node.Queries.ConvertTxId
  - Cardano.Node.Queries.KES
  - Cardano.Node.Queries.Ledger
