+++ cardano-node/src/Cardano/Tracing/Render.hs
@@ -31,7 +31,6 @@ import qualified Data.Text.Encoding as Text
 
 import           Cardano.BM.Tracing (TracingVerbosity (..))
 import           Cardano.Slotting.Slot (EpochNo (..), SlotNo (..), WithOrigin (..))
-import           Cardano.Tracing.ConvertTxId (ConvertTxId (..))
 import           Ouroboros.Consensus.Block (BlockNo (..), ConvertRawHash (..), RealPoint (..))
 import           Ouroboros.Consensus.Block.Abstract (Point (..))
 import           Ouroboros.Consensus.Ledger.SupportsMempool (GenTx, TxId)
@@ -39,6 +38,8 @@ import           Ouroboros.Consensus.Storage.ImmutableDB.Chunks.Internal (ChunkN
 import           Ouroboros.Consensus.Storage.ImmutableDB.Impl.Types (BlockOrEBB (..))
 import qualified Ouroboros.Consensus.Storage.ImmutableDB.API as ImmDB
 import           Ouroboros.Network.Block (ChainHash (..), HeaderHash, StandardHash, Tip, getTipPoint)
+
+import           Cardano.Node.Queries (ConvertTxId (..))
 
 renderBlockOrEBB :: BlockOrEBB -> Text
 renderBlockOrEBB (Block slotNo) = "Block at " <> renderSlotNo slotNo
